Port from tap to luatest
